requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Serialization
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Task Queue
celery==5.4.0
redis==5.2.1

# Object Storage
minio==7.2.12

# Elasticsearch
elasticsearch==8.17.0

# Vector Database
pymilvus==2.5.0

# LLM & AI
langchain==0.3.14
langchain-community==0.3.14
openai==1.58.1

# File Processing
python-docx==1.1.2
PyPDF2==3.0.1
openpyxl==3.1.5
Pillow==11.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
